factor io/ingest.py, create io/{_common,outgest}.py
